Refactor resource-related API to stream-based (#58)
* refactor(core): make resource process streaming

* refactor(core): remove file size limitation of Int.MAX_VALUE

* chore(core): arrange code

* fix(core): correctly handle source close

* refactor(milky): move Bot injection to MilkyContext init block

* refactor: make `resolveUri` return MediaSource

* refactor: make MediaSource abstract

* feat: implement contextual, disposable scope

* update(core): optimize ByteArrayMediaSource

* refactor: use streaming API in Codec and command execution

* docs: add warning for ffmpegPath config

* feat: add more rich logging for MediaSource

+++ acidify-core/src/commonMain/kotlin/org/ntqqrev/acidify/internal/proto/message/Ptt.kt
@@ -10,7 +10,7 @@ internal class Ptt(
     @ProtoNumber(3) val fileUuid: ByteArray = byteArrayOf(),
     @ProtoNumber(4) val fileMd5: ByteArray = byteArrayOf(),
     @ProtoNumber(5) val fileName: String = "",
-    @ProtoNumber(6) val fileSize: Int = 0,
+    @ProtoNumber(6) val fileSize: Long = 0L,
     @ProtoNumber(7) val reserve: ByteArray = byteArrayOf(),
     @ProtoNumber(8) val fileId: Int = 0,
     @ProtoNumber(9) val serverIp: Int = 0,
